+++ src/components/DurationInput/DurationInput.stories.tsx
@@ -2,7 +2,8 @@
 import React from 'react'
 import type { Meta, Story } from '@storybook/react'
 import { Title, Subtitle, ArgsTable, Stories, PRIMARY_STORY, Primary } from '@storybook/addon-docs/blocks'
-import { DurationInput, DurationInputProps, Layout, Label, TextInput } from '../..'
+import { Layout, Label, TextInput } from '../..'
+import { DurationInput, DurationInputProps, timeToDisplayText, DurationInputHelpers } from './DurationInput'
 
 export default {
   title: 'Form / DurationInput',
@@ -72,5 +73,36 @@ export const UsageRetainingStringInputDurationAndSupportsExpressionValue: Story<
       onChange={time => console.log(time)}
       {...args}
     />
+  )
+}
+export const Basic2: Story<DurationInputProps> = args => <DurationInput {...args} />
+
+Basic2.args = {
+  value: 93600000
+}
+
+export const Utils: Story = () => {
+  const [time, setTime] = React.useState(9600000)
+
+  function handleChange(e: React.ChangeEvent<HTMLInputElement>) {
+    const value = parseInt(e.target.value, 10)
+    setTime(value)
+  }
+
+  return (
+    <div>
+      <div style={{ marginBottom: '10px' }}>
+        <code>time</code>: <input type="number" value={time} onChange={handleChange} />
+      </div>
+      <div>
+        <code>timeToDisplayText(time)</code>: {timeToDisplayText(time)}
+      </div>
+      <div>
+        <code>DurationInputHelpers.VALID_SYNTAX_REGEX.test(timeToDisplayText(time))</code>:{' '}
+        {DurationInputHelpers.VALID_SYNTAX_REGEX.test(timeToDisplayText(time)) ? 'true' : 'false'}
+      </div>
+      <div></div>
+      <div></div>
+    </div>
   )
 }